# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FlowApp API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BookingStatus"
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get full booking detail (client, package, team, addons, invoice, files)
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
    patch:
      operationId: updateBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      responses:
        "204":
          description: Deleted
  /bookings/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /bookings/{bookingId}/files:
    parameters:
      - name: bookingId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listBookingFiles
      tags: [files]
      responses:
        "200":
          description: List of files for a booking
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeliveryFile"
    post:
      operationId: createBookingFile
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryFileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryFile"
  /files/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: deleteBookingFile
      tags: [files]
      responses:
        "204":
          description: Deleted
  /files/{id}/select:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: toggleFileSelection
      tags: [files]
      summary: Toggle client photo selection for a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSelectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryFile"
components:
  schemas:
...
    Package:
      type: object
      required: [id, name, price, includedEditedPhotos, estimatedDays, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        includedEditedPhotos:
          type: integer
        estimatedDays:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    AddOn:
      type: object
      required: [id, name, price, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        createdAt:
          type: string
          format: date-time
...
    TeamRole:
      type: string
      enum: [photographer, videographer, mua, hair_stylist, editor]
    TeamMember:
      type: object
      required: [id, name, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          $ref: "#/components/schemas/TeamRole"
        photoUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    ClientOrigin:
      type: string
      enum: [local, out_of_city, out_of_island, international]
    Client:
      type: object
      required: [id, name, email, whatsapp, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        clientOrigin:
          $ref: "#/components/schemas/ClientOrigin"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    BookingStatus:
      type: string
      enum: [pending, dp_paid, fully_paid, shooting, editing, delivered, closed]
    Booking:
      type: object
      required:
        [
          id,
          clientId,
          clientName,
          categoryId,
          eventDate,
          status,
          totalAmount,
          clientOrigin,
          createdAt,
        ]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        categoryId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        packageName:
          type: ["string", "null"]
        eventDate:
          type: string
          format: date-time
        locationName:
          type: ["string", "null"]
        locationAddress:
          type: ["string", "null"]
        mapsLink:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/BookingStatus"
        totalAmount:
          type: number
        clientOrigin:
          $ref: "#/components/schemas/ClientOrigin"
        specialRequest:
          type: ["string", "null"]
        googleDriveLink:
          type: ["string", "null"]
        detectSubfolder:
          type: boolean
        whatsappClient:
          type: ["string", "null"]
        whatsappAdmin:
          type: ["string", "null"]
        maxPhotos:
          type: integer
        pilihFotoEnabled:
          type: boolean
        downloadFotoEnabled:
          type: boolean
        pilihFotoDuration:
          type: string
        downloadFotoDuration:
          type: string
        pilihFotoPassword:
          type: ["string", "null"]
        downloadFotoPassword:
          type: ["string", "null"]
        pilihFotoTambahanEnabled:
          type: boolean
        pilihFotoCetakEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BookingInput:
      type: object
      required: [clientId, eventDate]
      properties:
        clientId:
          type: integer
        categoryId:
          type: integer
        packageId:
          type: integer
        eventDate:
          type: string
          format: date-time
        locationName:
          type: string
        locationAddress:
          type: string
        mapsLink:
          type: string
        clientOrigin:
          $ref: "#/components/schemas/ClientOrigin"
        specialRequest:
          type: string
        moodboardLinks:
          type: array
          items:
            type: string
        teamMemberIds:
          type: array
          items:
            type: integer
        addOnIds:
          type: array
          items:
            type: integer
        googleDriveLink:
          type: string
        detectSubfolder:
          type: boolean
        whatsappClient:
          type: string
        whatsappAdmin:
          type: string
        maxPhotos:
          type: integer
        pilihFotoEnabled:
          type: boolean
        downloadFotoEnabled:
          type: boolean
        pilihFotoDuration:
          type: string
        downloadFotoDuration:
          type: string
        pilihFotoPassword:
          type: string
        downloadFotoPassword:
          type: string
        pilihFotoTambahanEnabled:
          type: boolean
        pilihFotoCetakEnabled:
          type: boolean
    BookingUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        packageId:
          type: integer
        eventDate:
          type: string
          format: date-time
        locationName:
          type: string
        locationAddress:
          type: string
        mapsLink:
          type: string
        clientOrigin:
          $ref: "#/components/schemas/ClientOrigin"
        specialRequest:
          type: string
        moodboardLinks:
          type: array
          items:
            type: string
        teamMemberIds:
          type: array
          items:
            type: integer
        addOnIds:
          type: array
          items:
            type: integer
        googleDriveLink:
          type: string
        detectSubfolder:
          type: boolean
        whatsappClient:
          type: string
        whatsappAdmin:
          type: string
        maxPhotos:
          type: integer
        pilihFotoEnabled:
          type: boolean
        downloadFotoEnabled:
          type: boolean
        pilihFotoDuration:
          type: string
        downloadFotoDuration:
          type: string
        pilihFotoPassword:
          type: string
        downloadFotoPassword:
          type: string
        pilihFotoTambahanEnabled:
          type: boolean
        pilihFotoCetakEnabled:
          type: boolean
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/BookingStatus"
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          required: [client, package, teamMembers, addOns, moodboardLinks, invoice, files]
          properties:
            client:
              $ref: "#/components/schemas/Client"
            package:
              $ref: "#/components/schemas/Package"
            teamMembers:
              type: array
              items:
                $ref: "#/components/schemas/TeamMember"
            addOns:
              type: array
              items:
                $ref: "#/components/schemas/AddOn"
            moodboardLinks:
              type: array
              items:
                type: string
            invoice:
              oneOf:
                - $ref: "#/components/schemas/Invoice"
                - type: "null"
            files:
              type: array
              items:
                $ref: "#/components/schemas/DeliveryFile"
...
    InvoiceStatus:
      type: string
      enum: [unpaid, partial, paid]
...
    Invoice:
      type: object
      required: [id, bookingId, invoiceNumber, clientName, issueDate, dueDate, subtotal, total, paidAmount, status]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        invoiceNumber:
          type: string
        clientName:
          type: string
        issueDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        subtotal:
          type: number
        total:
          type: number
        paidAmount:
          type: number
        status:
          $ref: "#/components/schemas/InvoiceStatus"
...
    FolderType:
      type: string
      enum: [raw, edited, video, final_video, album]
    DeliveryFile:
      type: object
      required: [id, bookingId, folderType, fileName, fileUrl, selected, uploadedAt]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        folderType:
          $ref: "#/components/schemas/FolderType"
        fileName:
          type: string
        fileUrl:
          type: string
        selected:
          type: boolean
        uploadedAt:
          type: string
          format: date-time
    DeliveryFileInput:
      type: object
      required: [folderType, fileName, fileUrl]
      properties:
        folderType:
          $ref: "#/components/schemas/FolderType"
        fileName:
          type: string
          minLength: 1
        fileUrl:
          type: string
          minLength: 1
    FileSelectionUpdate:
      type: object
      required: [selected]
      properties:
        selected:
          type: boolean